Renamed compiler filenames to be clean and consistent
